Make some tests be integration tests
