+++ gedbrowser/src/test/java/org/schoellerfamily/gedbrowser/controller/test/LivingControllerIT.java
@@ -23,7 +23,9 @@ import org.springframework.test.web.servlet.client.RestTestClient;
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = {"management.port=0"})
 @AutoConfigureRestTestClient
-public class SubmittersControllerTest implements MenuTestHelper {
+@SuppressWarnings({ "PMD.JUnitTestsShouldIncludeAssert" })
+class LivingControllerIT implements MenuTestHelper {
+
     /**
      * Not sure what this is good for.
      */
@@ -38,9 +40,9 @@ public class SubmittersControllerTest implements MenuTestHelper {
 
     /** */
     @Test
-    void testSubmittersControllerOK() {
+    void testLivingControllerOK() {
         final String url = "http://localhost:" + port
-                + "/gedbrowser/submitters?db=gl120368";
+                + "/gedbrowser/living?db=gl120368";
         final EntityExchangeResult<String> entity = restTestClient.get()
                 .uri(URI.create(url))
                 .exchange()
@@ -50,31 +52,15 @@ public class SubmittersControllerTest implements MenuTestHelper {
             .returns(HttpStatus.OK.value(), result -> result.getStatus().value())
             .extracting(EntityExchangeResult::getResponseBody)
                 .asString().contains(
-                    "<title>Submitters - gl120368</title>",
-                    "Submitters for dataset: gl120368</h2>",
-                    "href=\"submitter?db=gl120368&amp;id=U1\">Phil Williams",
-                    "href=\"submitter?db=gl120368&amp;id=U3\"> ",
-                    "href=\"submitter?db=gl120368&amp;id=U14\">A K Robinson",
-                    "href=\"submitter?db=gl120368&amp;id=U15\">Michael ",
-                    "href=\"submitter?db=gl120368&amp;id=U2\">Arthur PUNCHA",
-                    "href=\"submitter?db=gl120368&amp;id=U4\">Created by Fa",
-                    "href=\"submitter?db=gl120368&amp;id=U5\">Paul Alger",
-                    "href=\"submitter?db=gl120368&amp;id=U6\">Jim Beecroft",
-                    "href=\"submitter?db=gl120368&amp;id=U7\">ken stel",
-                    "href=\"submitter?db=gl120368&amp;id=U8\">Mark Willis B",
-                    "href=\"submitter?db=gl120368&amp;id=U9\">Patricia Fisc",
-                    "href=\"submitter?db=gl120368&amp;id=U10\">Hirt-Klooze",
-                    "href=\"submitter?db=gl120368&amp;id=U11\">Lester LeMay",
-                    "href=\"submitter?db=gl120368&amp;id=U12\">David A. Blo",
-                    "href=\"submitter?db=gl120368&amp;id=U13\">Dave Morris",
+                    "<title>Living - gl120368</title>",
                     getMenu("A"));
     }
 
     /** */
     @Test
-    void testSubmittersControllerBadDataSet() {
+    void testLivingControllerBadDataSet() {
         final EntityExchangeResult<String> entity = restTestClient.get()
-                .uri(URI.create("http://localhost:" + port + "/gedbrowser/submitters?db=XYZZY"))
+                .uri(URI.create("http://localhost:" + port + "/gedbrowser/living?db=XYZZY"))
                 .exchange()
                 .returnResult(String.class);
