+++ gedbrowser/src/test/java/org/schoellerfamily/gedbrowser/controller/test/LoginControllerIT.java
@@ -27,7 +27,7 @@ import org.springframework.web.client.RestTemplate;
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = {"management.port=0"})
 @AutoConfigureRestTestClient
-class LoginControllerTest {
+class LoginControllerIT {
 
     /**
      * Not sure what this is good for.
